+++ src/Network/Network/Models/AzureFirewallPolicy/PSAzureFirewallPolicy.cs
@@ -24,6 +24,8 @@ namespace Microsoft.Azure.Commands.Network.Models
     {
         public string ThreatIntelMode { get; set; }
 
+        public PSAzureFirewallPolicyThreatIntelWhitelist ThreatIntelWhitelist { get; set; }
+
         public Microsoft.Azure.Management.Network.Models.SubResource BasePolicy { get; set; }
 
         public string ProvisioningState { get; set; }
@@ -31,5 +33,6 @@ namespace Microsoft.Azure.Commands.Network.Models
         [JsonProperty("ruleCollectionGroups")]
         public List<Microsoft.Azure.Management.Network.Models.SubResource> RuleCollectionGroups { get; set; }
 
+        public PSAzureFirewallPolicyDnsSettings DnsSettings { get; set; }
     }
 }