+++ src/Network/Network/Properties/Resources.Designer.cs
@@ -160,6 +160,33 @@ namespace Microsoft.Azure.Commands.Network.Properties {
         }
         
         /// <summary>
+        ///   Looks up a localized string similar to privateEndpointConnections.
+        /// </summary>
+        internal static string ApplicationGatewayPrivateEndpointConnectionName {
+            get {
+                return ResourceManager.GetString("ApplicationGatewayPrivateEndpointConnectionName", resourceCulture);
+            }
+        }
+        
+        /// <summary>
+        ///   Looks up a localized string similar to privateLinkConfigurations.
+        /// </summary>
+        internal static string ApplicationGatewayPrivateLinkConfigurationName {
+            get {
+                return ResourceManager.GetString("ApplicationGatewayPrivateLinkConfigurationName", resourceCulture);
+            }
+        }
+        
+        /// <summary>
+        ///   Looks up a localized string similar to ipConfigurations.
+        /// </summary>
+        internal static string ApplicationGatewayPrivateLinkIpConfigurationName {
+            get {
+                return ResourceManager.GetString("ApplicationGatewayPrivateLinkIpConfigurationName", resourceCulture);
+            }
+        }
+        
+        /// <summary>
         ///   Looks up a localized string similar to probes.
         /// </summary>
         internal static string ApplicationGatewayProbeName {
@@ -529,6 +556,24 @@ namespace Microsoft.Azure.Commands.Network.Properties {
         }
         
         /// <summary>
+        ///   Looks up a localized string similar to Circuit Connection with Name &apos;{0}&apos; is already added..
+        /// </summary>
+        internal static string ExpressRouteCircuitConnectionAlreadyAdded {
+            get {
+                return ResourceManager.GetString("ExpressRouteCircuitConnectionAlreadyAdded", resourceCulture);
+            }
+        }
+        
+        /// <summary>
+        ///   Looks up a localized string similar to Circuit Connection with Name &apos;{0}&apos; was not added to the private peering..
+        /// </summary>
+        internal static string ExpressRouteCircuitConnectionNotFound {
+            get {
+                return ResourceManager.GetString("ExpressRouteCircuitConnectionNotFound", resourceCulture);
+            }
+        }
+        
+        /// <summary>
         ///   Looks up a localized string similar to The ExpressRouteConnection {0} could not be found..
         /// </summary>
         internal static string ExpressRouteConnectionNotFound {
@@ -554,41 +599,16 @@ namespace Microsoft.Azure.Commands.Network.Properties {
                 return ResourceManager.GetString("ExpressRouteGatewayRequiredToCreateExpressRouteConnection", resourceCulture);
             }
         }
-
+        
         /// <summary>
-        ///   Looks up a localized string similar to A valid ExpressRouteGateway reference is required to create an ExpressRouteConnection..
+        ///   Looks up a localized string similar to Private Peering needs to be configured on the Express Route Circuit..
         /// </summary>
-        internal static string ExpressRoutePrivatePeeringNotFound
-        {
-            get
-            {
+        internal static string ExpressRoutePrivatePeeringNotFound {
+            get {
                 return ResourceManager.GetString("ExpressRoutePrivatePeeringNotFound", resourceCulture);
             }
         }
-
-        /// <summary>
-        ///   Looks up a localized string similar to A valid ExpressRouteGateway reference is required to create an ExpressRouteConnection..
-        /// </summary>
-        internal static string ExpressRouteCircuitConnectionNotFound
-        {
-            get
-            {
-                return ResourceManager.GetString("ExpressRouteCircuitConnectionNotFound", resourceCulture);
-            }
-        }
-
-
-        /// <summary>
-        ///   Looks up a localized string similar to A valid ExpressRouteGateway reference is required to create an ExpressRouteConnection..
-        /// </summary>
-        internal static string ExpressRouteCircuitConnectionAlreadyAdded
-        {
-            get
-            {
-                return ResourceManager.GetString("ExpressRouteCircuitConnectionAlreadyAdded", resourceCulture);
-            }
-        }
-
+        
         /// <summary>
         ///   Looks up a localized string similar to The hub virtual network connection to modify could not be found..
         /// </summary>